eCommerce_Task12-Adding Contact US Page
Adding Contact US Page via REACT ROUTER. Added a separate Navigator on the top called CONTACT US. When user clicks on it, this will take the user to the contact US page. In the Contact Us page, show the User a form which takes the following inputs -> Name, email Id, phone number of the user. When the user clicks on submit store the data into firebase by doing a POST request.